pyproject.toml
[project]
name = "veritarc-ai"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "black>=23.0.0",
    "datasets>=2.14.0",
    "faiss-cpu>=1.7.0",
    "fastapi>=0.100.0",
    "flake8>=6.0.0",
    "isort>=5.12.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langgraph>=0.0.20",
    "langsmith>=0.0.80",
    "loguru>=0.7.0",
    "numpy>=1.24.0",
    "openai>=1.0.0",
    "pandas>=2.0.0",
    "pillow>=10.0.0",
    "pydantic>=2.0.0",
    "pypdf2>=3.0.0",
    "pytesseract>=0.3.10",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "python-docx>=0.8.11",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "qdrant-client>=1.15.1",
    "ragas>=0.1.0",
    "rapidfuzz>=3.13.0",
    "requests>=2.31.0",
    "streamlit>=1.28.0",
    "tavily-python>=0.3.0",
    "tqdm>=4.65.0",
    "uvicorn[standard]>=0.23.0",
]
